Added Exception and Logger with Validation

+++ src/main/java/com/cg/onlinepizza/service/impl/CustomerServiceImpl.java
@@ -6,11 +6,15 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.cg.onlinepizza.dto.CustomerDTO;
 import com.cg.onlinepizza.dto.OrderDTO;
+import com.cg.onlinepizza.exceptions.CustomerAlreadyExistsException;
+import com.cg.onlinepizza.exceptions.CustomerNotFoundException;
+import com.cg.onlinepizza.exceptions.CustomersNotPresentException;
 import com.cg.onlinepizza.model.Customer;
 import com.cg.onlinepizza.model.Order;
 import com.cg.onlinepizza.repository.CustomerRepository;
@@ -19,10 +23,10 @@ import com.cg.onlinepizza.service.CustomerService;
 @Service
 public class CustomerServiceImpl implements CustomerService{
 
+	static Logger log = Logger.getLogger(CustomerServiceImpl.class.getName());
 	@Autowired
 	private CustomerRepository customerRepository;
 
-	@SuppressWarnings("unused")
 	public static CustomerDTO entityToDTO(Customer customer) 
 	{
 
@@ -51,7 +55,6 @@ public class CustomerServiceImpl implements CustomerService{
 		return customerDTO;
 	}
 
-	@SuppressWarnings("unused")
 	public static Customer DTOToEntity(CustomerDTO customerDTO)
 	{
 		Customer customer = new Customer();
@@ -63,50 +66,64 @@ public class CustomerServiceImpl implements CustomerService{
 		customer.setUserName(customerDTO.getUserName());
 		customer.setPassword(customerDTO.getPassword());
 		customer.setCustomerEmail(customerDTO.getCustomerEmail());
-
-		//		Set<Order> order = new HashSet<>();
-		//		Set<OrderDTO> orderDTO = new HashSet<>(customerDTO.getOrder());
-		//		if(orderDTO == null)
-		//		{
-		//			customer.setOrder(null);
-		//		}
-		//		else
-		//		{
-		//			for(OrderDTO o: orderDTO)
-		//			{
-		//				order.add(OrderServiceImpl.DTOToEntity(o));
-		//			}
-		//
-		//			customer.setOrder(order);
-		//		}
 		customer.setOrder(null);
 		return customer;
 	}
 
 	@Override
-	public List<CustomerDTO> getAllCustomers() {
+	public List<CustomerDTO> getAllCustomers() throws CustomersNotPresentException {
+		
+		log.info("Service Layer - Entry - get Customers");
+		
 		List<CustomerDTO> customerDTOReturn = new ArrayList<>();
 		for(Customer c:customerRepository.findAll())
 		{
 			customerDTOReturn.add(entityToDTO(c));
 		}
-
+        if(customerDTOReturn.isEmpty())
+        {
+        	throw new CustomersNotPresentException("No Customers Available");
+        }
+        
+        log.info("Service Layer - Exit - get Customers");
+        
 		return customerDTOReturn;
 	}
 
 	@Override
-	public List<CustomerDTO> deleteCustomer(Integer customerId) {
+	public List<CustomerDTO> deleteCustomer(Integer customerId) throws CustomerNotFoundException{
+		
+		log.info("Service Layer - Entry - delete Customers");
+		
+		Optional<Customer> checking = customerRepository.findById(customerId);
+		if(checking.isEmpty())
+		{
+			 throw new CustomerNotFoundException("Customer Not Found To Delete"); 
+		}
 		customerRepository.deleteById(customerId);
 		List<CustomerDTO> customerDTOReturn = new ArrayList<>();
 		for(Customer c:customerRepository.findAll())
 		{
+			log.warn("WARN: delete customers Started");
+			
 			customerDTOReturn.add(entityToDTO(c));
 		}
+		
+		log.info("Service Layer - Exit - delete Customers");
+		
 		return customerDTOReturn;
 	}
 
 	@Override
-	public List<CustomerDTO> saveCustomer(CustomerDTO customerDTO) {
+	public List<CustomerDTO> saveCustomer(CustomerDTO customerDTO) throws CustomerAlreadyExistsException {
+		
+		log.info("Service Layer - Entry - save Customers");
+		
+		Optional<Customer> checking = customerRepository.findById(DTOToEntity(customerDTO).getCustomerId());
+		if(checking.isPresent())
+		{
+			throw new CustomerAlreadyExistsException("Customer Already Exists");
+		}
 		Customer customer = new Customer();
 		customer = DTOToEntity(customerDTO);
 
@@ -117,11 +134,23 @@ public class CustomerServiceImpl implements CustomerService{
 		{
 			customerDTOReturn.add(entityToDTO(c));
 		}
+		
+		log.info("Service Layer - Exit - save Customers");
+		
 		return customerDTOReturn;
 	}
 
 	@Override
-	public List<CustomerDTO> updateCustomer(CustomerDTO customerDTO) {
+	public List<CustomerDTO> updateCustomer(CustomerDTO customerDTO) throws CustomerNotFoundException{
+		
+		log.info("Service Layer - Entry - update Customers");
+		
+		Optional<Customer>customers=customerRepository.findById(DTOToEntity(customerDTO).getCustomerId());
+		
+		  if(customers.isEmpty()) { 
+			  throw new CustomerNotFoundException("Customer Not Found"); 
+			  }
+		
 		Customer customer = new Customer();
 		customer = DTOToEntity(customerDTO);
 
@@ -132,13 +161,23 @@ public class CustomerServiceImpl implements CustomerService{
 		{
 			customerDTOReturn.add(entityToDTO(c));
 		}
-
+		log.info("Service Layer - Exit - Update Customers");
 		return  customerDTOReturn;
 	}
 
 	@Override
-	public CustomerDTO findCustomer(Integer customerId) {
+	public CustomerDTO findCustomer(Integer customerId) throws CustomerNotFoundException {
+		
+		log.info("Service Layer - Entry - find Customer");
+		
 		Optional<Customer>customer=customerRepository.findById(customerId);
+		
+		  if(customer.isEmpty()) { 
+			  
+			  throw new CustomerNotFoundException("Customer Not Found"); 
+			  }
+		
+		 log.info("Service Layer - Exit - find Customer");	
 		return entityToDTO(customer.get());
 	}
